Update PR after inital feedback
